Regen typeshare models (#189)
* Regen typeshare models

* remove generated

* re-gen models

* Update docker.yml

* remove extra comment

* Update docker.yml

* add CLAUDE.md

* Update docker.yml

* feat: add fallback to main branch when specified tag/branch not found in Docker build

+++ blockchain/src/main/kotlin/com/gemwallet/android/blockchain/clients/solana/SolanaStakeClient.kt
@@ -43,12 +43,12 @@ class SolanaStakeClient(
         delegations.map { delegation ->
             val info = delegation.account.data.parsed.info
             val deactivateEpoch = try {
-                info.stake.delegation.deactivationEpoch.toInt()
+                info.stake.delegation.deactivationEpoch.toLong()
             } catch (_: Throwable) {
                 null
             }
             val activationEpoch = try {
-                info.stake.delegation.activationEpoch.toInt()
+                info.stake.delegation.activationEpoch.toLong()
             } catch (_: Throwable) {
                 null
             }
